Opgave06 23/12/15 19:41

+++ Opgave06/BackendImplementation/Worker.cs
@@ -13,7 +13,7 @@ namespace BackendImplementation
     public class Worker : IWorker
     {
         public event Action<string> CollisionFound;
-        public event Action<ulong> ProgressChanged;
+        public event Action<decimal> ProgressChanged;
         public event Action Stalled;
         private string hashwoord;
